loader: Load from a directory representing a line
From now on a line will be defined in a directory containg
.layouts file (and other future extension i.e. .acei, .ql etc..)
For easy of use there can be subdirectories rappresenting portion
of the line (i.e. stations) containg all the relevant files.

The Loader class now traverse all the files and subdirectories under
the base line directory and loads all the .layout files.